refactor: convert pages to server components
