+++ apps/profile/components/home/home-animations.tsx
@@ -1,0 +1,17 @@
+"use client";
+
+import { useRef } from "react";
+import { useAnimateOnScroll } from "@/lib/animations";
+
+export function HomeAnimations({ children }: { children: React.ReactNode }) {
+  const containerRef = useRef(null);
+  useAnimateOnScroll(containerRef);
+
+  return (
+    <div ref={containerRef} className="mx-auto px-4 max-w-7xl">
+      {children}
+    </div>
+  );
+}
+
+HomeAnimations.displayName = "HomeAnimations";
